validation_service/requirements.txt
grpcio==1.68.0
grpcio-tools==1.68.0
protobuf==5.29.0
pika==1.3.1
requests==2.31.0
python-dotenv
